@@ -1,8 +1,10 @@
 package com.abitcreative.popularmovies.activites;
 
 import android.os.Bundle;
-import android.support.annotation.Nullable;
-import android.support.v7.app.AppCompatActivity;
+
+
+import androidx.annotation.Nullable;
+import androidx.appcompat.app.AppCompatActivity;
 
 import com.abitcreative.popularmovies.R;
 
