@@ -1,11 +1,12 @@
 package com.abitcreative.popularmovies.adapters;
 
 import android.content.Context;
-import android.support.v7.widget.RecyclerView;
+
 import android.view.LayoutInflater;
-import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ImageView;
+
+import androidx.recyclerview.widget.RecyclerView;
 
 import com.abitcreative.popularmovies.R;
 import com.abitcreative.popularmovies.webapi.ListResult;
