@@ -3,11 +3,12 @@ package com.abitcreative.popularmovies.adapters;
 import android.content.Context;
 import android.content.Intent;
 import android.net.Uri;
-import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.TextView;
+
+import androidx.recyclerview.widget.RecyclerView;
 
 import com.abitcreative.popularmovies.R;
 import com.abitcreative.popularmovies.webapi.Video;
